backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
pydantic>=2.0
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart